+++ packages/atomic/src/components/search/search-box-suggestions/atomic-search-box-instant-results/atomic-search-box-instant-results.tsx
@@ -22,7 +22,7 @@ import {
   ResultDisplayDensity,
   ResultDisplayImageSize,
   ResultDisplayLayout,
-} from '../../atomic-result/atomic-result-display-options';
+} from '../../../common/layout/display-options';
 import {Button} from '../../../common/button';
 
 /**
@@ -43,7 +43,7 @@ export class AtomicSearchBoxInstantResults implements BaseResultList {
 
   private results: Result[] = [];
   public resultListCommon!: ResultListCommon;
-  private renderingFunction: ResultRenderingFunction | null = null;
+  private renderingFunction?: ResultRenderingFunction | undefined;
 
   /**
    * Sets a rendering function to bypass the standard HTML template mechanism for rendering results.
@@ -55,7 +55,6 @@ export class AtomicSearchBoxInstantResults implements BaseResultList {
    */
   @Method() public async setRenderFunction(render: ResultRenderingFunction) {
     this.renderingFunction = render;
-    this.assignRenderingFunctionIfPossible();
   }
   /**
    * The maximum number of results to show.
@@ -112,6 +111,7 @@ export class AtomicSearchBoxInstantResults implements BaseResultList {
     const results = this.instantResults.state.results.length
       ? this.instantResults.state.results
       : this.results;
+
     const elements: SearchBoxSuggestionElement[] = results.map(
       (result: Result) => ({
         key: `instant-result-${encodeForDomAttribute(result.uniqueId)}`,
@@ -127,6 +127,7 @@ export class AtomicSearchBoxInstantResults implements BaseResultList {
             imageSize={this.imageSize}
             content={this.resultListCommon.getContentOfResultTemplate(result)}
             stopPropagation={false}
+            renderingFunction={this.renderingFunction}
           ></atomic-result>
         ),
         onSelect: (e: MouseEvent) => {
@@ -171,9 +172,7 @@ export class AtomicSearchBoxInstantResults implements BaseResultList {
       host: this.host,
       bindings: this.bindings,
       templateElements: this.host.querySelectorAll('atomic-result-template'),
-      onReady: () => {
-        this.assignRenderingFunctionIfPossible();
-      },
+      onReady: () => {},
       onError: () => {
         this.templateHasError = true;
       },
@@ -218,11 +217,6 @@ export class AtomicSearchBoxInstantResults implements BaseResultList {
     });
   }
 
-  private assignRenderingFunctionIfPossible() {
-    if (this.resultListCommon && this.renderingFunction) {
-      this.resultListCommon.renderingFunction = this.renderingFunction;
-    }
-  }
   public render() {
     if (this.error) {
       return (