+++ packages/quantic/force-app/main/default/lwc/quanticFacetValue/quanticFacetValue.js
@@ -11,6 +11,7 @@ import LOCALE from '@salesforce/i18n/locale';
  * The `QuanticFacetValue` component is used by a facet component to display a formatted facet value and the number of results with that value.
  * @fires CustomEvent#selectvalue
  * @category Search
+ * @category Insight Panel
  * @example
  * <c-quantic-facet-value onselectvalue={onSelect} item={result} is-checked={result.checked} display-as-link={displayAsLink} formatting-function={formattingFunction}></c-quantic-facet-value>
  */
@@ -81,6 +82,7 @@ export default class QuanticFacetValue extends LightningElement {
       })
     );
   }
+  
   /**
    * @param {KeyboardEvent} evt
    */